next

+++ Backend/src/routes/favRourtes.js
@@ -5,11 +5,15 @@ const Product = require("../Schema");
 const verifyToken = require("./verifyUser");
 
 router.get("/", verifyToken, async (req, res) => {
-  const fav = await Fav.find({ user: req.user._id });
-  if (fav[0]) {
-    res.send(fav[0].items);
-  } else {
-    res.send([]);
+  try {
+    const fav = await Fav.find({ user: req.user._id });
+    if (fav[0]) {
+      res.send(fav[0].items);
+    } else {
+      res.send([]);
+    }    
+  } catch (err) {
+    res.send({error:err})
   }
 });
 
@@ -17,13 +21,22 @@ router.post("/:id", verifyToken, async (req, res) => {
   const favourite = await Fav.find({ user: req.user._id });
   const prod = await Product.findById(req.params.id);
   if (favourite[0]) {
-    const hello = new Fav({
-      user: req.user._id,
-      items: [...favourite[0].items, { product: prod, img: req.body.img }],
-    });
-    await Fav.deleteOne(favourite._id);
-    await hello.save();
-    res.send({ message: "updated" });
+    var p = favourite[0].items.filter((p)=>p.product.toString()===req.params.id.toString());
+    if(p[0]){
+      res.send({message:"Already in fav"});
+    }else{
+      try {
+        const hello = new Fav({
+          user: req.user._id,
+          items: [...favourite[0].items, { product: prod, img: req.body.img }],
+        });
+        await Fav.deleteOne(favourite._id);
+        await hello.save();
+        res.send({ message: "updated" });      
+      } catch (err) {
+        res.send({error:err})
+      }
+  }
   } else {
     const createdFav = new Fav({
       user: req.user._id,
@@ -36,7 +49,6 @@ router.post("/:id", verifyToken, async (req, res) => {
 
 router.get("/:id", verifyToken, async (req, res) => {
   const favourite = await Fav.find({ user: req.user._id });
-  // console.log(favourite);
   try {
     const hello = new Fav({
       user: req.user._id,
@@ -46,10 +58,8 @@ router.get("/:id", verifyToken, async (req, res) => {
     });
     await Fav.deleteOne(favourite._id);
     await hello.save();
-    // console.log(hello);
     res.send(hello.items);
   } catch (err) {
-    // console.log(err);
     res.send({ error: err });
   }
 });
